backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.5
python-dateutil==2.9.0.post0
feedparser==6.0.11
beautifulsoup4==4.12.3
sentence-transformers==3.3.1
scikit-learn==1.6.1
